fix: fixes dataabase operations issues
At the paramsSchema, nonoptional() is not needed because if its not provided any nullable statement by default it infers required. Then at the UpdatePartyById function, i was comparing the argument id (that is referent to the party id) to the person_id attribute, making that the desired party wasn't updated. At the getAllPersonPartyById function the filter need to be in % sign so that the word passed as the filter can be at eny position in the party name.